# Improved generic int impl