--- a/starlark/int_generic.go
+++ b/starlark/int_generic.go
@@ -1,34 +1,26 @@
-//go:build (!linux && !darwin && !dragonfly && !freebsd && !netbsd && !solaris) || (!amd64 && !arm64 && !mips64x && !ppc64x && !loong64)
-// +build !linux,!darwin,!dragonfly,!freebsd,!netbsd,!solaris !amd64,!arm64,!mips64x,!ppc64x,!loong64
+//go:build (!linux && !darwin && !dragonfly && !freebsd && !netbsd && !solaris) || (!amd64 && !arm64 && !mips64x && !ppc64x && !loong64) || noposixint
+// +build !linux,!darwin,!dragonfly,!freebsd,!netbsd,!solaris !amd64,!arm64,!mips64x,!ppc64x,!loong64 noposixint
 
 package starlark
 
-// generic Int implementation as a union
-
 import "math/big"
 
-type intImpl struct {
-	// We use only the signed 32-bit range of small to ensure
-	// that small+small and small*small do not overflow.
-	small_ int64    // minint32 <= small <= maxint32
-	big_   *big.Int // big != nil <=> value is not representable as int32
-}
+const hasPosixInts = false
 
-// --- low-level accessors ---
-
-// get returns the small and big components of the Int.
-// small is defined only if big is nil.
-// small is sign-extended to 64 bits for ease of subsequent arithmetic.
-func (i Int) get() (small int64, big *big.Int) {
-	return i.impl.small_, i.impl.big_
-}
-
-// Precondition: math.MinInt32 <= x && x <= math.MaxInt32
-func makeSmallInt(x int64) Int {
-	return Int{intImpl{small_: x}}
+// int_get returns the (small, big) arms of the union.
+func int_get(i Int) (int64, *big.Int) {
+	switch i := i.(type) {
+	case intSmall:
+		return int64(i), nil
+	case *intBig:
+		return 0, (*big.Int)(i)
+	default:
+		panic("Int is not an int?")
+	}
 }
 
 // Precondition: x cannot be represented as int32.
-func makeBigInt(x *big.Int) Int {
-	return Int{intImpl{big_: x}}
-}
+func makeBigInt(x *big.Int) Int { return (*intBig)(x) }
+
+// Precondition: math.MinInt32 <= x && x <= math.MaxInt32
+func makeSmallInt(x int64) Int { return intSmall(x) }