In this PR I propose a new generic representation for Int(s) which would work for all non-posix non-x64 targets. 

Moreover, I propose to add a build tag to completely remove the integer optimizations even for targets like linux x64 since being able to allocate 4GB of ram is not enough of a reason to believe that consuming 4GB of virtual memory is irrelevant. 

For example:

```bash
# Running the test suite on a linux x64 box, running in the folder <repo>/starlark

# 2 GB, it works
( ulimit -v $(( 2 * 1 << 20 )) && go test . -count=1 ) 

# 4 GB, it works
( ulimit -v $(( 4 * 1 << 20 )) && go test . -count=1 ) 

# 5 GB, it crashes with OOM
( ulimit -v $(( 5 * 1 << 20 )) && go test . -count=1 ) 
```

IMHO, it should never happen that **adding** more than twice the memory causes an "Out Of Memory" for an otherwise working program.

## Benchmarking

Here follows some benchmarking, performed on an Ubuntu 22.10, with an Intel Core i5.

### x86 

Current master:
```
goos: linux
goarch: 386
pkg: go.starlark.net/starlark
cpu: Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz
BenchmarkStarlark/bench_bigint-4         	    1719	    914096 ns/op	   96120 B/op	    7008 allocs/op
BenchmarkStarlark/bench_gauss-4          	      25	  48124428 ns/op	 4637752 B/op	  316324 allocs/op
BenchmarkStarlark/bench_int-4            	    2826	    516878 ns/op	   32040 B/op	    2002 allocs/op
BenchmarkStarlark/bench_mix-4            	    2570	    466473 ns/op	   53404 B/op	    1961 allocs/op
PASS
ok  	go.starlark.net/starlark	9.426s
```

This PR:
```
goos: linux
goarch: 386
pkg: go.starlark.net/starlark
cpu: Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz
BenchmarkStarlark/bench_bigint-4         	    1869	    647316 ns/op	   70031 B/op	    5749 allocs/op
BenchmarkStarlark/bench_gauss-4          	      27	  43207239 ns/op	 2951792 B/op	  289580 allocs/op
BenchmarkStarlark/bench_int-4            	    2424	    506344 ns/op	   11943 B/op	    1490 allocs/op
BenchmarkStarlark/bench_mix-4            	    3367	    373223 ns/op	   32215 B/op	     636 allocs/op
PASS
ok  	go.starlark.net/starlark	6.826s
```

Both faster and consumes less memory.

### x64

#### With the optimization on (no memory limits)

Current master:
```
goos: linux
goarch: amd64
pkg: go.starlark.net/starlark
cpu: Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz
BenchmarkStarlark/bench_bigint-4         	    3208	    436642 ns/op	  128176 B/op	    5006 allocs/op
BenchmarkStarlark/bench_gauss-4          	      80	  18628541 ns/op	 3387445 B/op	  132323 allocs/op
BenchmarkStarlark/bench_int-4            	    9541	    119273 ns/op	      48 B/op	       1 allocs/op
BenchmarkStarlark/bench_mix-4            	    6494	    167812 ns/op	   63608 B/op	     636 allocs/op
PASS
ok  	go.starlark.net/starlark	7.744s
```

This PR:
```
goos: linux
goarch: amd64
pkg: go.starlark.net/starlark
cpu: Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz
BenchmarkStarlark/bench_bigint-4         	    2724	    422222 ns/op	  128176 B/op	    5006 allocs/op
BenchmarkStarlark/bench_gauss-4          	      75	  19292476 ns/op	 3387464 B/op	  132323 allocs/op
BenchmarkStarlark/bench_int-4            	    9775	    130073 ns/op	      48 B/op	       1 allocs/op
BenchmarkStarlark/bench_mix-4            	    6882	    179801 ns/op	   63608 B/op	     636 allocs/op
PASS
ok  	go.starlark.net/starlark	6.337s
```

The difference in the metrics seems varying from run to run, the allocation pattern is the same.

#### With memory limits (ulimit -v 2000000)

Current master:
```
2023/01/20 16:53:56 Starlark failed to allocate 4GB address space: cannot allocate memory. Integer performance may suffer.
goos: linux
goarch: amd64
pkg: go.starlark.net/starlark
cpu: Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz
BenchmarkStarlark/bench_bigint-4         	    1654	    640482 ns/op	  168210 B/op	    7007 allocs/op
BenchmarkStarlark/bench_gauss-4          	      36	  40755274 ns/op	 9688947 B/op	  447395 allocs/op
BenchmarkStarlark/bench_int-4            	    3549	    355176 ns/op	   80080 B/op	    4002 allocs/op
BenchmarkStarlark/bench_mix-4            	    3476	    387892 ns/op	  116208 B/op	    3236 allocs/op
PASS
ok  	go.starlark.net/starlark	7.836s
```

This PR:
```
2023/01/20 16:54:14 Starlark failed to allocate 4GB address space: cannot allocate memory. Integer performance may suffer.
goos: linux
goarch: amd64
pkg: go.starlark.net/starlark
cpu: Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz
BenchmarkStarlark/bench_bigint-4         	    2896	    401010 ns/op	  134104 B/op	    5749 allocs/op
BenchmarkStarlark/bench_gauss-4          	      52	  21606800 ns/op	 4645504 B/op	  289580 allocs/op
BenchmarkStarlark/bench_int-4            	    7246	    160877 ns/op	   11967 B/op	    1490 allocs/op
BenchmarkStarlark/bench_mix-4            	    6639	    213752 ns/op	   63616 B/op	     636 allocs/op
PASS
ok  	go.starlark.net/starlark	7.837s
```

Both faster and consumes less memory. Small ints suffer more than big ones as it must allocate values above 255.