@adonovan 2023-01-21
Thanks for reporting this issue. The underlying problem here is that TestProfile allocates too much (around 2GB), so that if the virtual address space has been artificially limited to 5GB, then there isn't enough left after the optimization to represent the actual memory allocated by the program. I've reported this at #455.

We can't change the public API of Int even if we wanted to, but fortunately it isn't necessary. The appropriate fix is to improve TestProfile, and perhaps also to add an option (e.g. a flag or env var) to allow users to disable the int optimization entirely.


> being able to allocate 4GB of ram is not enough of a reason to believe that consuming 4GB of virtual memory is irrelevant.
> it should never happen that adding more than twice the memory causes an "Out Of Memory" for an otherwise working program.

To be clear, the optimization never attempts to allocate 4GB of RAM (though I think you're aware of that). Are virtual address spaces limited to 5GB actually a problem in practice? 4GB, yes (e.g. 386) ; 700MB, yes (e.g. Android). I'm curious to know how you encountered this problem in production.