@marco6 2023-01-23
> The underlying problem here is that TestProfile allocates too much (around 2GB)

The underlying problem is that as soon as you have enough virtual space to "book" 4GB, then immediately the available address space drops, which is rather strange from a usage POV. This would be true for any usage of Starlark, not just `TestProfile`.  While 2 GB for a single Starlark execution seems a lot, this problem would arise in any project using Starlark as a scripting back-end as in that case there could be multiple (possibly a lot) running scripts at the same time, easily consuming gigabytes of memory.

> We can't change the public API of Int

I understand, and I agree 100%, so, for me, we can close this PR. Maybe a issue is better suited for this kind of discussion?

> To be clear, the optimization never attempts to allocate 4GB of RAM (though I think you're aware of that).

Yeah sure. I guess I don't know a better verb other than "allocate" to say "booked a range in the virtual address space", which is rather confusing in this situation.

> Are virtual address spaces limited to 5GB actually a problem in practice?

Not in general, but with it could be, as you saw with other platforms. I still think that this behavior should be tunable at compile-time or at run-time as it can create problems (it did!) and we don't really know under which condition this piece of code will run.

I think I could summarize the issue I'm facing with:

  - `TestIntFallback` is right now broken on any platform that doesn't support that optimization (just run ` GOARCH=386 go test .`: it fails for no reason);
  - on a supported platform, as soon you have, during startup, enough address space to book 4GB, your available address space will drop. Note that **this depends on the package initialization order** (which depends on the other modules of your project and is not under the control of the programmer[^1], you only get a "loose" (topological) ordering, if I read the specs right);
  - if the "allocation" (or mapping? or booking? whatever...) fails, the implementation doesn't switch back to the one you would get on an x64 one, but it will use an even worse one (all `big.Int`s), so we have 3 behaviors: the "best" one, the "generic" one (compile-time), and the worse one (run-time, when the allocation fails).
 
[^1]: oh well as per [this](https://github.com/golang/go/issues/31636) thread, it actually **is** under the control of the programmer, using and undocumented feature of `cgo` only. On the other side, [this](https://groups.google.com/g/golang-nuts/c/-1-40xkpirg) other discussion seems to be different. Confusing eh? Still doesn't feel like the right thing to do.