@adonovan 2023-01-28
> The underlying problem is that as soon as you have enough virtual space to "book" 4GB, then immediately the available address space drops, which is rather strange from a usage POV.

I agree it may be surprising, but I wonder whether and how this situation actually occurs in practice. Generally, the limit on the address space is either too small for the optimization (e.g. 4GB or 700MB) or unimaginably vast (2^64 or so). I haven't seen much in the middle, and I can't yet think of any reason why someone would artificially limit their address space (as opposed to max heap size) to a value close to the amount of memory needed by the program. So I'd like to understand this before we try to make a fix.

I'd be happy to add an environment variable to disable the optimization, but of course you'd still have to hit the problem at least once to know to set the var. But that seems fine.


> This would be true for any usage of Starlark, not just TestProfile. While 2 GB for a single Starlark execution seems a lot, this problem would arise in any project using Starlark as a scripting back-end as in that case there could be multiple (possibly a lot) running scripts at the same time, easily consuming gigabytes of memory.

If each script or interpreter is running in its own process, then there's no problem, as each process has an independent ulimit on its address space. 


> I understand, and I agree 100%, so, for me, we can close this PR. Maybe a issue is better suited for this kind of discussion?

An issue is fine, but discussing it here seems fine too.


> TestIntFallback is right now broken on any platform that doesn't support that optimization (just run  GOARCH=386 go test .: it fails for no reason); on a supported platform, as soon you have, during startup, enough address space to book 4GB, your available address space will drop.

I haven't tested it on 386 recently, but we can certainly tweak (or disable) the tests to make it more portable or less fragile. 

> Note that this depends on the package initialization order 

I'm not worried about initialization order: as you say, it's a topological order, and it is well documented by the spec as a predictable and deterministic function of a given source tree, so I don't think it's an interesting source of unpleasant surprises in practice.